Agregado el codigo de verificacion de variable de sesion

+++ admin/eliminarcliente.php
@@ -1,9 +1,14 @@
 <?php 
+session_start();
+if(!isset($_SESSION['usuario'])) 
+{
+  header('Location: login.php'); 
+  exit();
+}
 include ("../config/config.php");
 $id=$conn->QSTR($_GET['id'],get_magic_quotes_gpc());	
 $consulta=$conn->Execute("DELETE FROM clientes WHERE id=$id");	
 ?>
-
 <script type="text/javascript">
 window.location="clientes.php";
 </script>
